cleanup more xcode warnings

+++ SignalServiceKit/src/Messages/Attachments/OWSThumbnailService.swift
@@ -49,11 +49,11 @@ private struct OWSThumbnailRequest {
     public typealias FailureBlock = (Error) -> Void
 
     let attachment: TSAttachmentStream
-    let thumbnailDimensionPoints: UInt
+    let thumbnailDimensionPoints: CGFloat
     let success: SuccessBlock
     let failure: FailureBlock
 
-    init(attachment: TSAttachmentStream, thumbnailDimensionPoints: UInt, success: @escaping SuccessBlock, failure: @escaping FailureBlock) {
+    init(attachment: TSAttachmentStream, thumbnailDimensionPoints: CGFloat, success: @escaping SuccessBlock, failure: @escaping FailureBlock) {
         self.attachment = attachment
         self.thumbnailDimensionPoints = thumbnailDimensionPoints
         self.success = success
@@ -98,7 +98,7 @@ public class OWSThumbnailService: NSObject {
     // success and failure will be called async _off_ the main thread.
     @objc
     public func ensureThumbnail(forAttachment attachment: TSAttachmentStream,
-                                thumbnailDimensionPoints: UInt,
+                                thumbnailDimensionPoints: CGFloat,
                                 success: @escaping SuccessBlock,
                                 failure: @escaping FailureBlock) {
         serialQueue.async {
